+++ app/src/main/java/com/keylesspalace/tusky/entity/Attachment.kt
@@ -58,4 +58,24 @@ data class Attachment(
             }
         }
     }
+
+    /**
+     * The meta data of an [Attachment].
+     */
+    @Parcelize
+    data class MetaData (
+            val focus: Focus?
+    ) : Parcelable
+
+    /**
+     * The Focus entity, used to specify the focal point of an image.
+     *
+     * See here for more details what the x and y mean:
+     *   https://github.com/jonom/jquery-focuspoint#1-calculate-your-images-focus-point
+     */
+    @Parcelize
+    data class Focus (
+            val x: Float,
+            val y: Float
+    ) : Parcelable
 }
