--- app/src/main/java/com/keylesspalace/tusky/entity/MetaData.kt
@@ -1,28 +1,0 @@
-/* Copyright 2017 Jochem Raat <jchmrt@riseup.net>
- *
- * This file is a part of Tusky.
- *
- * This program is free software; you can redistribute it and/or modify it under the terms of the
- * GNU General Public License as published by the Free Software Foundation; either version 3 of the
- * License, or (at your option) any later version.
- *
- * Tusky is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
- * Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with Tusky; if not,
- * see <http://www.gnu.org/licenses>. */
-
-package com.keylesspalace.tusky.entity
-
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-
-/**
- * The meta data of an [Attachment].
- */
-@Parcelize
-data class MetaData (
-        // The focus attribute of the meta-data, if it is specified.
-        val focus: Focus?
-) : Parcelable
